Extract Development.UnitTests.UnitTesting.Sandboxing to own project
