+++ Remotion.sln
@@ -372,6 +372,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Web.IntegrationTests", "Rem
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ObjectBinding.Web.IntegrationTests", "Remotion\ObjectBinding\Web.IntegrationTests\ObjectBinding.Web.IntegrationTests.csproj", "{963B16B7-01CE-4AF0-940E-B8D7B086856F}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Development.Sandboxing.NUnit2", "Remotion\Development\Sandboxing.NUnit2\Development.Sandboxing.NUnit2.csproj", "{8E50108A-309F-4933-B054-D44D6065DA80}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Development.Sandboxing.NUnit2.UnitTests", "Remotion\Development\Sandboxing.NUnit2.UnitTests\Development.Sandboxing.NUnit2.UnitTests.csproj", "{34C50163-EFE8-45E1-9D8C-B2DCC8A00573}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -1392,6 +1396,22 @@ Global
 		{963B16B7-01CE-4AF0-940E-B8D7B086856F}.Release|Any CPU.Build.0 = Release|Any CPU
 		{963B16B7-01CE-4AF0-940E-B8D7B086856F}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
 		{963B16B7-01CE-4AF0-940E-B8D7B086856F}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{8E50108A-309F-4933-B054-D44D6065DA80}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8E50108A-309F-4933-B054-D44D6065DA80}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8E50108A-309F-4933-B054-D44D6065DA80}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{8E50108A-309F-4933-B054-D44D6065DA80}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{8E50108A-309F-4933-B054-D44D6065DA80}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8E50108A-309F-4933-B054-D44D6065DA80}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8E50108A-309F-4933-B054-D44D6065DA80}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{8E50108A-309F-4933-B054-D44D6065DA80}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{34C50163-EFE8-45E1-9D8C-B2DCC8A00573}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{34C50163-EFE8-45E1-9D8C-B2DCC8A00573}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{34C50163-EFE8-45E1-9D8C-B2DCC8A00573}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{34C50163-EFE8-45E1-9D8C-B2DCC8A00573}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{34C50163-EFE8-45E1-9D8C-B2DCC8A00573}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{34C50163-EFE8-45E1-9D8C-B2DCC8A00573}.Release|Any CPU.Build.0 = Release|Any CPU
+		{34C50163-EFE8-45E1-9D8C-B2DCC8A00573}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{34C50163-EFE8-45E1-9D8C-B2DCC8A00573}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -1524,6 +1544,8 @@ Global
 		{73119503-AEC8-4256-9E6C-ABABB1F1EC6F} = {72195525-BF6C-43D6-AC96-39E34D02EA34}
 		{6800F4E0-363D-4FF9-8B25-4A916D164867} = {2E32EB1E-B3AA-4ACF-BA20-EC6FD44CC05B}
 		{963B16B7-01CE-4AF0-940E-B8D7B086856F} = {72195525-BF6C-43D6-AC96-39E34D02EA34}
+		{8E50108A-309F-4933-B054-D44D6065DA80} = {9F4D90EC-6AD2-495D-BFC3-270DC15974DD}
+		{34C50163-EFE8-45E1-9D8C-B2DCC8A00573} = {9F4D90EC-6AD2-495D-BFC3-270DC15974DD}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {26742782-2BDC-4959-9951-6D45B54A5C0D}
