+++ Remotion/Development/Sandboxing.NUnit2.UnitTests/UnitTesting/DummyTest4.cs
@@ -17,30 +17,30 @@
 using System;
 using NUnit.Framework;
 
-namespace Remotion.Development.UnitTests.Core.UnitTesting.Sandboxing
+namespace Remotion.Development.Sandboxing.NUnit2.UnitTests.UnitTesting
 {
   [TestFixture]
-  public class DummyTest3
+  public class DummyTest4
   {
-    private int _value;
+    private int _value = Math.Abs (-5);
 
-    [SetUp]
-    public void SetUp ()
+    [TearDown]
+    public void TearDown ()
     {
-      _value = 10;
+      _value = Math.Abs (-5);
     }
 
     [Test]
     public void Test1 ()
     {
-      Assert.That (_value, Is.EqualTo (10));
+      Assert.That (_value, Is.EqualTo (5));
       _value = 100;
     }
 
     [Test]
     public void Test2 ()
     {
-      Assert.That (_value, Is.EqualTo (10));
+      Assert.That (_value, Is.EqualTo (5));
       _value = 200;
     }
   }
