+++ Remotion/Development/Sandboxing.NUnit2.UnitTests/UnitTesting/DummyTestIgnored.cs
@@ -16,9 +16,9 @@
 // 
 using System;
 using NUnit.Framework;
-using Remotion.Development.UnitTesting.Sandboxing;
+using Remotion.Development.Sandboxing.Nunit2.UnitTesting;
 
-namespace Remotion.Development.UnitTests.Core.UnitTesting.Sandboxing
+namespace Remotion.Development.Sandboxing.NUnit2.UnitTests.UnitTesting
 {
   [Ignore]
   public class DummyTestIgnored
