+++ Remotion/Development/Sandboxing.NUnit2.UnitTests/UnitTesting/PermissionSetsTest.cs
@@ -23,9 +23,9 @@ using System.Net.Mail;
 using System.Security.Permissions;
 using System.Web;
 using NUnit.Framework;
-using Remotion.Development.UnitTesting.Sandboxing;
+using Remotion.Development.Sandboxing.Nunit2.UnitTesting;
 
-namespace Remotion.Development.UnitTests.Core.UnitTesting.Sandboxing
+namespace Remotion.Development.Sandboxing.NUnit2.UnitTests.UnitTesting
 {
   [TestFixture]
   public class PermissionSetsTest
