+++ Remotion/Development/Sandboxing.NUnit2.UnitTests/UnitTesting/SandboxTestRunnerTest.cs
@@ -17,9 +17,9 @@
 using System;
 using System.Linq;
 using NUnit.Framework;
-using Remotion.Development.UnitTesting.Sandboxing;
+using Remotion.Development.Sandboxing.Nunit2.UnitTesting;
 
-namespace Remotion.Development.UnitTests.Core.UnitTesting.Sandboxing
+namespace Remotion.Development.Sandboxing.NUnit2.UnitTests.UnitTesting
 {
   [TestFixture]
   public class SandboxTestRunnerTest
