+++ Remotion/Development/Sandboxing.NUnit2/UnitTesting/TestFixtureResult.cs
@@ -19,7 +19,7 @@ using System;
 // Note: This file is originally defined in Remotion.Development.UnitTesting.Sandboxing. It is duplicated by Remotion.Linq.UnitTests.Sandboxing.
 // Note: Changes made to this file must be synchronized with all copies.
 
-namespace Remotion.Development.UnitTesting.Sandboxing
+namespace Remotion.Development.Sandboxing.Nunit2.UnitTesting
 {
   /// <summary>
   /// <see cref="TestFixtureResult"/> holds the type of the test class and the result of the test methods.
